Work towards compiling cal3d 64-bit clean (allow gnu license: modifying our exporters)
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@111741 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ plugins/src/SkeletonCandidate.cpp
@@ -304,7 +304,7 @@ bool CSkeletonCandidate::CreateFromSkeletonFile(const std::string& strFilename)
 	std::vector<boost::shared_ptr<CalCoreBone> >& vectorCoreBone = m_skeleton->coreBones;
 
 	// loop through all root core bones
-	std::vector<int>::const_iterator iteratorRootCoreBoneId;
+	std::vector<size_t>::const_iterator iteratorRootCoreBoneId;
 	for(iteratorRootCoreBoneId = m_skeleton->rootBoneIds.begin(); iteratorRootCoreBoneId != m_skeleton->rootBoneIds.end(); ++iteratorRootCoreBoneId)
 	{
 		// recursively add the core bone to the skeleton candidate
